Per issue #81 - Moving logic into the scheduler.
