+++ src/MDMScheduler.h
@@ -18,6 +18,7 @@
 
 @protocol MDMSchedulerDelegate;
 @protocol MDMPlan;
+@protocol MDMNamedPlan;
 
 /**
  The possible activity states a scheduler can be in.
@@ -62,6 +63,28 @@ NS_SWIFT_NAME(Scheduler)
 /** Associate a plan with a given target. */
 - (void)addPlan:(nonnull NSObject<MDMPlan> *)plan toTarget:(nonnull id)target
     NS_SWIFT_NAME(addPlan(_:to:));
+
+/**
+ Associates a named plan with a given target.
+ 
+ @param plan The plan to add to this transaction.
+ @param name String identifier for the plan.
+ @param target The target on which the plan can operate.
+ */
+- (void)addPlan:(nonnull id<MDMNamedPlan>)plan
+          named:(nonnull NSString *)name
+       toTarget:(nonnull id)target
+NS_SWIFT_NAME(addPlan(_:named:to:));
+
+/**
+ Removes any plan associated with the given name on the given target.
+ 
+ @param name String identifier for the plan.
+ @param target The target on which the plan can operate.
+ */
+- (void)removePlanNamed:(nonnull NSString *)name
+             fromTarget:(nonnull id)target
+NS_SWIFT_NAME(removePlan(named:from:));
 
 #pragma mark Committing transactions
